backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
starlette==0.37.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.2
email-validator==2.3.0
httpx==0.28.1
